+++ assembler/src/projects/dipspades/dipspades_config.hpp
@@ -14,7 +14,7 @@ struct dipspades_config {
 
     struct base_params {
         size_t             K;
-        size_t             max_threads;
+        unsigned           max_threads;
         size_t             max_memory;
         size_t             read_buffer_size;
     };
